requirements.txt
llama-index
openai
python-dotenv
python-pptx
requests